+++ src/FluentDbTools/Abstractions/FluentDbTools.Common.Abstractions/SupportedDatabaseTypes.cs
@@ -1,8 +1,17 @@
 namespace FluentDbTools.Common.Abstractions
 {
+    /// <summary>
+    /// Contains Currently supported database types
+    /// </summary>
     public enum SupportedDatabaseTypes
     {
+        /// <summary>
+        /// Support for Postgres Database
+        /// </summary>
         Postgres,
+        /// <summary>
+        /// Support for Oracle Database
+        /// </summary>
         Oracle
     }
 }